WIP: 0 bounce works for small volumes

+++ src/pbrt/util/parallel.h
@@ -266,7 +266,7 @@ void ParallelFor2D(const Bounds2i &extent, std::function<void(Bounds2i)> func);
 // Parallel Inline Functions
 inline void ParallelFor(int64_t start, int64_t end, bool runInParallel, std::function<void(int64_t)> func) {
     if (!runInParallel)
-        for (int i = start; i < end; ++i)
+        for (int64_t i = start; i < end; ++i)
             func(i);
 
     else {
